Packaging (#83)
* Adding tests for image property issue manager

* Added test for image_property_issue_manager.set_params()

* Added test for image_property_issue_manager._get_defer_set()

* Changed structure of init method to be more consistent

* Fixed tests

* Added tests for methods in DuplicateIssueManager

* Added test for DuplicateIssueManager._get_issue_types_to_compute

* Added unit tests for duplicate issue manager

* Fixed bugs after tests

* Updated comments

* black formatting

* Updated documentation for test_update_info in test_duplicate_issue_manager.py

* Improved ids

* Updated test_remove_exact_duplicates_from_near

* Removed redundant test cases

* Changed constants to all caps

* Added a before check for set_params test

* Added ids for tests in TestImagePropertyIssueManager

* added before check for set_params() test

* Added save and load functionality

* Adding documentation on how to use cleanvision

* Added detailed instructions on how to use Imagelab

* Removed unnecessary file committed

* Apply suggestions from code review


* Update cleanvision/imagelab.py


* Update cleanvision/imagelab.py


* Update cleanvision/imagelab.py


* Added check for invalid path and empty directory

* Removed duplicate line

* Updated demo notebook with Ulyana's comments

* Added pyproject.toml config for build

* Fixed missing dependency

* Specified filepath as a command line argument for example script

* Updated path in jupyter notebook

* Fixed CI pipeline

---------
