+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include LICENSE README.md CONTRIBUTING.md DEVELOPMENT.md CODE_OF_CONDUCT.md
+include CONTRIBUTING.md DEVELOPMENT.md CODE_OF_CONDUCT.md
