+++ pyproject.toml
@@ -1,0 +1,46 @@
+[build-system]
+requires = ["setuptools>=43.0.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "cleanvision"
+version = "0.0.1"
+authors = [
+    { name = "Cleanlab Inc.", email = "team@cleanlab.ai" },
+]
+description = "Find issues in image datasets"
+readme = "README.md"
+requires-python = ">=3.7"
+keywords = ["computer_vision", "cv", "image_data", "issue_detection", "data_quality", "image_quality", "machine_learning", "data_cleaning", "image_deduplication"]
+license = { file = "LICENSE" }
+classifiers = [
+    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.7",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python",
+    "Topic :: Software Development",
+    "Topic :: Scientific/Engineering",
+    "Topic :: Scientific/Engineering :: Mathematics",
+    "Topic :: Scientific/Engineering :: Artificial Intelligence",
+    "Topic :: Software Development :: Libraries",
+    "Topic :: Software Development :: Libraries :: Python Modules"
+]
+dependencies = [
+    "Pillow>=9.3",
+    "numpy>=1.20.0",
+    "pandas>=1.1.5",
+    "tabulate>=0.8.3", # pandas optional dependency for .to_markdown()
+    "imagehash>=4.2.0",
+    "tqdm>=4.53.0",
+    "matplotlib>=3.4"
+]
+
+[project.urls]
+"Source" = "https://github.com/cleanlab/cleanvision"
+"Bug Tracker" = "https://github.com/cleanlab/cleanvision/issues"
+#todo add documentation link
+
