+++ src/cleanvision/utils/constants.py
@@ -1,0 +1,14 @@
+from typing import List
+
+IMAGE_PROPERTY: str = "image_property"
+DUPLICATE: str = "duplicate"
+IMAGE_PROPERTY_ISSUE_TYPES_LIST: List[str] = [
+    "dark",
+    "light",
+    "odd_aspect_ratio",
+    "low_information",
+    "blurry",
+    "grayscale",
+]
+DUPLICATE_ISSUE_TYPES_LIST: List[str] = ["exact_duplicates", "near_duplicates"]
+SETS: str = "sets"
